+++ input.c
@@ -1605,6 +1605,9 @@ key_press_release(struct seat *seat, struct terminal *term, uint32_t serial,
     if (released)
         stop_repeater(seat, key);
 
+    if (pressed)
+        seat->kbd.last_key_press_was_shortcut = false;
+
     bool should_repeat =
         pressed && xkb_keymap_key_repeats(seat->kbd.xkb_keymap, key);
 
@@ -1706,6 +1709,7 @@ key_press_release(struct seat *seat, struct terminal *term, uint32_t serial,
                 if (bind->k.sym == raw_syms[i] &&
                     execute_binding(seat, term, bind, serial, 1))
                 {
+                    seat->kbd.last_key_press_was_shortcut = true;
                     goto maybe_repeat;
                 }
             }
@@ -1719,6 +1723,7 @@ key_press_release(struct seat *seat, struct terminal *term, uint32_t serial,
                 bind->mods == (mods & ~consumed) &&
                 execute_binding(seat, term, bind, serial, 1))
             {
+                seat->kbd.last_key_press_was_shortcut = true;
                 goto maybe_repeat;
             }
         }
@@ -1734,10 +1739,28 @@ key_press_release(struct seat *seat, struct terminal *term, uint32_t serial,
                 if (code->item == key &&
                     execute_binding(seat, term, bind, serial, 1))
                 {
+                    seat->kbd.last_key_press_was_shortcut = true;
                     goto maybe_repeat;
                 }
             }
         }
+    }
+
+    if (released && seat->kbd.last_key_press_was_shortcut) {
+        /*
+         * Don't process a release event, if it corresponds to a
+         * triggered shortcut.
+         *
+         * 1. If we consumed a key (press) event, we shouldn't emit an
+         *    escape for its release event.
+         * 2. Ignoring the incorrectness of doing so; this also caused
+         *    us to reset the viewport.
+         *
+         * Background: if the kitty keyboard protocol was enabled,
+         * then the viewport was instantly reset to the bottom, after
+         * scrolling up.
+         */
+        goto maybe_repeat;
     }
 
     /*